Clase 12/11/2025

+++ backend/controller/tareasControllers.js
@@ -2,7 +2,7 @@ const asyncHandler = require('express-async-handler');
 const Tarea = require('../models/tareasModel');
 
 const getTareas = asyncHandler(async(req, res) => {
-    const tareas = await Tarea.find();
+    const tareas = await Tarea.find({user: req.user.id});
     res.status(200).json(tareas);
 })
 
@@ -12,7 +12,8 @@ const crearTareas = asyncHandler(async(req, res) => {
         throw new Error("Por favor agrega un texto para la tarea");
     }
     const tarea = await Tarea.create({
-        texto: req.body.texto
+        texto: req.body.texto,
+        user: req.user.id
 
     });
     res.status(201).json(tarea);
@@ -25,8 +26,16 @@ const updateTareas = asyncHandler(async(req, res) => {
         res.status(404);
         throw new Error("Tarea no encontrada");
     }
-    const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, {new: true});
-    res.status(200).json(tareaUpdated);
+
+    //Verificar si la tarea es del usuario Logueado :)
+    if(tarea.user.toString() !== req.user.id){
+        res.status(401);
+        throw new Error("Acceso no autorizado")
+    } else{
+        const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, {new: true});
+        res.status(200).json(tareaUpdated);
+    }
+
 })
 
 const deleteTareas = asyncHandler(async(req, res) => {
@@ -36,8 +45,17 @@ const deleteTareas = asyncHandler(async(req, res) => {
         res.status(404);
         throw new Error("Tarea no encontrada");
     }
-    await tarea.deleteOne();
-    res.status(200).json({id: req.params.id});
+    
+    //Verificar si la tarea es del usuario Logueado :)
+    if(tarea.user.toString() !== req.user.id){
+        res.status(401);
+        throw new Error("Acceso no autorizado")
+    } else{
+        await tarea.deleteOne();
+        res.status(200).json({id: req.params.id});
+    }
+
+
 })
 
 module.exports = {
